Refactor backend into modular routers
